aisi_inference/docker_files/requirements.txt
tqdm==4.61.2
torchvision==0.10.0
requests==2.25.1
torch==1.9.0
pycocotools==2.0.2
uvicorn==0.15.0
numpy==1.21.0
fastapi==0.68.1
Pillow==9.1.0
scikit_learn==0.24.2
python-multipart==0.0.5
opencv-python==4.5.5.64
opencv-contrib-python==4.5.5.64
bbox==0.9.2